Rename component to be consistent
