+++ ui/src/flux/components/BodyBuilder.tsx
@@ -3,7 +3,7 @@ import _ from 'lodash'
 
 import FancyScrollbar from 'src/shared/components/FancyScrollbar'
 import ExpressionNode from 'src/flux/components/ExpressionNode'
-import VariableName from 'src/flux/components/VariableName'
+import VariableNode from 'src/flux/components/VariableNode'
 import FuncSelector from 'src/flux/components/FuncSelector'
 import BodyDelete from 'src/flux/components/BodyDelete'
 import {funcNames} from 'src/flux/constants'
@@ -33,7 +33,7 @@ class BodyBuilder extends PureComponent<Props> {
                 {!dIndex && (
                   <BodyDelete bodyID={b.id} onDeleteBody={onDeleteBody} />
                 )}
-                <VariableName name={d.name} assignedToQuery={true} />
+                <VariableNode name={d.name} assignedToQuery={true} />
                 <ExpressionNode
                   bodyID={b.id}
                   declarationID={d.id}
@@ -49,7 +49,7 @@ class BodyBuilder extends PureComponent<Props> {
           return (
             <div className="declaration" key={i}>
               <div className="func-node--wrapper">
-                <VariableName name={b.source} assignedToQuery={false} />
+                <VariableNode name={b.source} assignedToQuery={false} />
                 <div className="func-node--menu">
                   <BodyDelete
                     bodyID={b.id}
